requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.0
faiss-cpu==1.11.0
fastapi==0.115.12
filelock==3.13.1
fsspec==2024.6.1
h11==0.16.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
pandas==2.2.3
pillow==11.0.0
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
torch==2.7.0+cpu
torchaudio==2.7.0+cpu
torchvision==0.22.0+cpu
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
uvicorn==0.34.2
